Add string parsing functionality for text boxes
The new TextDelimiters script can be used to define characters that can
be inserted into text files (e.g. for dialogue boxes) to be replaced
with other characters in-game. Currently this is only used to insert
new lines into text boxes (as a new line in the text file creates a new
paragraph in-game), but functionality can be expanded.

TODO It's rather messily done now, should make it easier to add in new
delimiters.